requirements.txt
Flask==2.3.3
python-dotenv==1.0.0
google-generativeai==0.1.0
Pillow==10.0.0